Backend/requirements.txt
Django==5.1.7
djangorestframework==3.14.0
psycopg[binary]
django-cors-headers==4.3.1
gunicorn==21.2.0
whitenoise==6.6.0
requests==2.31.0
djangorestframework-simplejwt==5.3.0
Pillow==11.3.0
django-ratelimit==4.1.0
python-decouple==3.8
boto3==1.34.0 